rename TransportLayer class to TransportNode
